Create review controller
