+++ app/controllers/candidate_interface/course_choices/replace_choices/base_controller.rb
@@ -29,34 +29,6 @@ module CandidateInterface
           end
         end
 
-        def confirm_choice
-          @course_choice = current_application.application_choices.find(params['id'])
-          @replacement_course_option = CourseOption.find(params['course_option_id'])
-        end
-
-        def update_choice
-          @course_choice = current_application.application_choices.find(params['id'])
-          @replacement_course_option_id = params['course_option_id']
-
-          @pick_site = PickSiteForm.new(
-            application_form: current_application,
-            provider_id: @course_choice.provider.id,
-            course_id: @course_choice.course.id,
-            study_mode: @course_choice.course_option.study_mode,
-            course_option_id: @replacement_course_option_id,
-          )
-
-          if @pick_site.valid?
-            @course_choice.update!(course_option_id: @replacement_course_option_id)
-            flash[:success] = 'Your application has been updated'
-
-            redirect_to candidate_interface_application_complete_path
-          else
-            flash[:warning] = 'Please select a new location.'
-            redirect_to candidate_interface_replace_course_choice_location_path(@course_choice.id)
-          end
-        end
-
       private
 
         def render_404_if_flag_is_inactive
