+++ app/views/candidate_interface/course_choices/replace_choices/review/confirm_choice.html.erb
