+++ config/routes.rb
@@ -206,8 +206,8 @@ Rails.application.routes.draw do
         post '/replace/:id' => 'course_choices/replace_choices/decision#route_action'
         get '/replace/:id/location' => 'course_choices/replace_choices/site_selection#replace_location', as: :replace_course_choice_location
         post '/replace/:id/location' => 'course_choices/replace_choices/site_selection#validate_location', as: :validate_new_course_choice_location
-        get '/replace/:id/confirm/:course_option_id' => 'course_choices/replace_choices/base#confirm_choice', as: :confirm_replacement_course_choice
-        get '/replace/:id/update/:course_option_id' => 'course_choices/replace_choices/base#update_choice', as: :update_replacement_course_choice
+        get '/replace/:id/confirm/:course_option_id' => 'course_choices/replace_choices/review#confirm_choice', as: :confirm_replacement_course_choice
+        get '/replace/:id/update/:course_option_id' => 'course_choices/replace_choices/review#update_choice', as: :update_replacement_course_choice
 
         get '/provider' => 'course_choices/provider_selection#new', as: :course_choices_provider
         post '/provider' => 'course_choices/provider_selection#create'
